+++ types/react-event-listener/index.d.ts
@@ -2,228 +2,224 @@
 // Project: https://github.com/oliviertassinari/react-event-listener
 // Definitions by: Alexey Svetliakov <https://github.com/asvetliakov>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
-// TypeScript Version: 2.8
 
 import React = require("react");
 
-export interface EventOptions {
-    /**
-     * @default false
-     */
-    passive?: boolean;
-    /**
-     * @default false
-     */
-    capture?: boolean;
-}
+export interface EventOptions extends Pick<AddEventListenerOptions, 'capture' | 'passive'> { }
 
-export function withOptions<T>(handler: (ev: T) => any, options: EventOptions): (ev: T) => any;
+export function withOptions<T, TThis = any>(handler: (this: TThis, ev: T) => any, options: EventOptions): (this: TThis, ev: T) => any;
 
-export interface EventListenerProps {
+export type EventListenerThisType<T extends EventTarget | WindowEventTargets> = T extends keyof Window ? Window[T] : T;
+
+export type OnErrorEventHandlerArgs = OnErrorEventHandlerNonNull extends (...args: infer A) => any ? A : [];
+
+export interface EventListenerProps<T extends EventTarget | WindowEventTargets> {
     // Global events
-    onPointerCancel?(ev: PointerEvent): any;
-    onPointerCancelCapture?(ev: PointerEvent): any;
-    onPointerDown?(ev: PointerEvent): any;
-    onPointerDownCapture?(ev: PointerEvent): any;
-    onPointerEnter?(ev: PointerEvent): any;
-    onPointerEnterCapture?(ev: PointerEvent): any;
-    onPointerLeave?(ev: PointerEvent): any;
-    onPointerLeaveCapture?(ev: PointerEvent): any;
-    onPointerMove?(ev: PointerEvent): any;
-    onPointerMoveCapture?(ev: PointerEvent): any;
-    onPointerOut?(ev: PointerEvent): any;
-    onPointerOutCapture?(ev: PointerEvent): any;
-    onPointerOver?(ev: PointerEvent): any;
-    onPointerOverCapture?(ev: PointerEvent): any;
-    onPointerUp?(ev: PointerEvent): any;
-    onPointerUpCapture?(ev: PointerEvent): any;
-    onWheel?(ev: WheelEvent): any;
-    onWheelCapture?(ev: WheelEvent): any;
-    onAbort?(ev: Event): any;
-    onAbortCapture?(ev: Event): any;
-    onAfterPrint?(ev: Event): any;
-    onAfterPrintCapture?(ev: Event): any;
-    onBeforePrint?(ev: Event): any;
-    onBeforePrintCapture?(ev: Event): any;
-    onBeforeUnload?(ev: BeforeUnloadEvent): any;
-    onBeforeUnloadCapture?(ev: BeforeUnloadEvent): any;
-    onBlur?(ev: FocusEvent): any;
-    onBlurCapture?(ev: FocusEvent): any;
-    onCanPlay?(ev: Event): any;
-    onCanPlayCapture?(ev: Event): any;
-    onCanPlayThrough?(ev: Event): any;
-    onCanPlayThroughCapture?(ev: Event): any;
-    onChange?(ev: Event): any;
-    onChangeCapture?(ev: Event): any;
-    onClick?(ev: MouseEvent): any;
-    onClickCapture?(ev: MouseEvent): any;
-    onCompassNeedsCalibration?(ev: Event): any;
-    onCompassNeedsCalibrationCapture?(ev: Event): any;
-    onContextMenu?(ev: PointerEvent): any;
-    onContextMenuCapture?(ev: PointerEvent): any;
-    onDblClick?(ev: MouseEvent): any;
-    onDblClickCapture?(ev: MouseEvent): any;
-    onDeviceMotion?(ev: DeviceMotionEvent): any;
-    onDeviceMotionCapture?(ev: DeviceMotionEvent): any;
-    onDeviceOrientation?(ev: DeviceOrientationEvent): any;
-    onDeviceOrientationCapture?(ev: DeviceOrientationEvent): any;
-    onDrag?(ev: DragEvent): any;
-    onDragCapture?(ev: DragEvent): any;
-    onDragEnd?(ev: DragEvent): any;
-    onDragEndCapture?(ev: DragEvent): any;
-    onDragEnter?(ev: DragEvent): any;
-    onDragEnterCapture?(ev: DragEvent): any;
-    onDragLeave?(ev: DragEvent): any;
-    onDragLeaveCapture?(ev: DragEvent): any;
-    onDragOver?(ev: DragEvent): any;
-    onDragOverCapture?(ev: DragEvent): any;
-    onDragStart?(ev: DragEvent): any;
-    onDragStartCapture?(ev: DragEvent): any;
-    onDrop?(ev: DragEvent): any;
-    onDropCapture?(ev: DragEvent): any;
-    onDurationChange?(ev: Event): any;
-    onDurationChangeCapture?(ev: Event): any;
-    onEmptied?(ev: Event): any;
-    onEmptiedCapture?(ev: Event): any;
-    onEnded?(ev: Event): any;
-    onEndedCapture?(ev: Event): any;
-    onError?: ErrorEventHandler;
-    onErrorCapture?: ErrorEventHandler;
-    onFocus?(ev: FocusEvent): any;
-    onFocusCapture?(ev: FocusEvent): any;
-    onHashChange?(ev: HashChangeEvent): any;
-    onHashChangeCapture?(ev: HashChangeEvent): any;
-    onInput?(ev: Event): any;
-    onInputCapture?(ev: Event): any;
-    onKeyDown?(ev: KeyboardEvent): any;
-    onKeyDownCapture?(ev: KeyboardEvent): any;
-    onKeyPress?(ev: KeyboardEvent): any;
-    onKeyPressCapture?(ev: KeyboardEvent): any;
-    onKeyUp?(ev: KeyboardEvent): any;
-    onKeyUpCapture?(ev: KeyboardEvent): any;
-    onLoad?(ev: Event): any;
-    onLoadCapture?(ev: Event): any;
-    onLoadedData?(ev: Event): any;
-    onLoadedDataCapture?(ev: Event): any;
-    onLoadedMetadata?(ev: Event): any;
-    onLoadedMetadataCapture?(ev: Event): any;
-    onLoadStart?(ev: Event): any;
-    onLoadStartCapture?(ev: Event): any;
-    onMessage?(ev: MessageEvent): any;
-    onMessageCapture?(ev: MessageEvent): any;
-    onMouseDown?(ev: MouseEvent): any;
-    onMouseDownCapture?(ev: MouseEvent): any;
-    onMouseEnter?(ev: MouseEvent): any;
-    onMouseEnterCapture?(ev: MouseEvent): any;
-    onMouseLeave?(ev: MouseEvent): any;
-    onMouseLeaveCapture?(ev: MouseEvent): any;
-    onMouseMove?(ev: MouseEvent): any;
-    onMouseMoveCapture?(ev: MouseEvent): any;
-    onMouseOut?(ev: MouseEvent): any;
-    onMouseOutCapture?(ev: MouseEvent): any;
-    onMouseOver?(ev: MouseEvent): any;
-    onMouseOverCapture?(ev: MouseEvent): any;
-    onMouseUp?(ev: MouseEvent): any;
-    onMouseUpCapture?(ev: MouseEvent): any;
-    onMouseWheel?(ev: WheelEvent): any;
-    onMouseWheelCapture?(ev: WheelEvent): any;
-    onMsGestureChange?(ev: MSGestureEvent): any;
-    onMsGestureChangeCapture?(ev: MSGestureEvent): any;
-    onMsGestureDoubleTap?(ev: MSGestureEvent): any;
-    onMsGestureDoubleTapCapture?(ev: MSGestureEvent): any;
-    onMsGestureEnd?(ev: MSGestureEvent): any;
-    onMsGestureEndCapture?(ev: MSGestureEvent): any;
-    onMsGestureHold?(ev: MSGestureEvent): any;
-    onMsGestureHoldCapture?(ev: MSGestureEvent): any;
-    onMsGestureStart?(ev: MSGestureEvent): any;
-    onMsGestureStartCapture?(ev: MSGestureEvent): any;
-    onMsGestureTap?(ev: MSGestureEvent): any;
-    onMsGestureTapCapture?(ev: MSGestureEvent): any;
-    onMsInertiaStart?(ev: MSGestureEvent): any;
-    onMsInertiaStartCapture?(ev: MSGestureEvent): any;
-    onMsPointerCancel?(ev: MSPointerEvent): any;
-    onMsPointerCancelCapture?(ev: MSPointerEvent): any;
-    onMsPointerDown?(ev: MSPointerEvent): any;
-    onMsPointerDownCapture?(ev: MSPointerEvent): any;
-    onMsPointerEnter?(ev: MSPointerEvent): any;
-    onMsPointerEnterCapture?(ev: MSPointerEvent): any;
-    onMsPointerLeave?(ev: MSPointerEvent): any;
-    onMsPointerLeaveCapture?(ev: MSPointerEvent): any;
-    onMsPointerMove?(ev: MSPointerEvent): any;
-    onMsPointerMoveCapture?(ev: MSPointerEvent): any;
-    onMsPointerOut?(ev: MSPointerEvent): any;
-    onMsPointerOutCapture?(ev: MSPointerEvent): any;
-    onMsPointerOver?(ev: MSPointerEvent): any;
-    onMsPointerOverCapture?(ev: MSPointerEvent): any;
-    oNmsPointerUp?(ev: MSPointerEvent): any;
-    oNmsPointerUpCapture?(ev: MSPointerEvent): any;
-    onOffline?(ev: Event): any;
-    onOfflineCapture?(ev: Event): any;
-    onOnline?(ev: Event): any;
-    onOnlineCapture?(ev: Event): any;
-    onOrientationChange?(ev: Event): any;
-    onOrientationChangeCapture?(ev: Event): any;
-    onPageHide?(ev: PageTransitionEvent): any;
-    onPageHideCapture?(ev: PageTransitionEvent): any;
-    onPageShow?(ev: PageTransitionEvent): any;
-    onPageShowCapture?(ev: PageTransitionEvent): any;
-    onPause?(ev: Event): any;
-    onPauseCapture?(ev: Event): any;
-    onPlay?(ev: Event): any;
-    onPlayCapture?(ev: Event): any;
-    onPlaying?(ev: Event): any;
-    onPlayingCapture?(ev: Event): any;
-    onPopState?(ev: PopStateEvent): any;
-    onPopStateCapture?(ev: PopStateEvent): any;
-    onProgress?(ev: ProgressEvent): any;
-    onProgressCapture?(ev: ProgressEvent): any;
-    onRateChange?(ev: Event): any;
-    onRateChangeCapture?(ev: Event): any;
-    onReadyStateChange?(ev: ProgressEvent): any;
-    onReadyStateChangeCapture?(ev: ProgressEvent): any;
-    onReset?(ev: Event): any;
-    onResetCapture?(ev: Event): any;
-    onResize?(ev: UIEvent): any;
-    onResizeCapture?(ev: UIEvent): any;
-    onScroll?(ev: UIEvent): any;
-    onScrollCapture?(ev: UIEvent): any;
-    onSeeked?(ev: Event): any;
-    onSeekedCapture?(ev: Event): any;
-    onSeeking?(ev: Event): any;
-    onSeekingCapture?(ev: Event): any;
-    onSelect?(ev: UIEvent): any;
-    onSelectCapture?(ev: UIEvent): any;
-    onSelectionChange?(ev: Event): any;
-    onSelectionChangeCapture?(ev: Event): any;
-    onStalled?(ev: Event): any;
-    onStalledCapture?(ev: Event): any;
-    onStorage?(ev: StorageEvent): any;
-    onStorageCapture?(ev: StorageEvent): any;
-    onSubmit?(ev: Event): any;
-    onSubmitCapture?(ev: Event): any;
-    onSuspend?(ev: Event): any;
-    onSuspendCapture?(ev: Event): any;
-    onTimeUpdate?(ev: Event): any;
-    onTimeUpdateCapture?(ev: Event): any;
-    onTouchCancel?: any;
-    onTouchCancelCapture?: any;
-    onTouchEnd?: any;
-    onTouchEndCapture?: any;
-    onTouchMove?: any;
-    onTouchMoveCapture?: any;
-    onTouchStart?: any;
-    onTouchStartCapture?: any;
-    onUnload?(ev: Event): any;
-    onUnloadCapture?(ev: Event): any;
-    onVolumeChange?(ev: Event): any;
-    onVolumeChangeCapture?(ev: Event): any;
-    onWaiting?(ev: Event): any;
-    onWaitingCapture?(ev: Event): any;
+    onPointerCancel?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerCancelCapture?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerDown?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerDownCapture?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerEnter?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerEnterCapture?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerLeave?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerLeaveCapture?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerMove?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerMoveCapture?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerOut?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerOutCapture?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerOver?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerOverCapture?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerUp?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onPointerUpCapture?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onWheel?(this: EventListenerThisType<T>, ev: WheelEvent): any;
+    onWheelCapture?(this: EventListenerThisType<T>, ev: WheelEvent): any;
+    onAbort?(this: EventListenerThisType<T>, ev: Event): any;
+    onAbortCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onAfterPrint?(this: EventListenerThisType<T>, ev: Event): any;
+    onAfterPrintCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onBeforePrint?(this: EventListenerThisType<T>, ev: Event): any;
+    onBeforePrintCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onBeforeUnload?(this: EventListenerThisType<T>, ev: BeforeUnloadEvent): any;
+    onBeforeUnloadCapture?(this: EventListenerThisType<T>, ev: BeforeUnloadEvent): any;
+    onBlur?(this: EventListenerThisType<T>, ev: FocusEvent): any;
+    onBlurCapture?(this: EventListenerThisType<T>, ev: FocusEvent): any;
+    onCanPlay?(this: EventListenerThisType<T>, ev: Event): any;
+    onCanPlayCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onCanPlayThrough?(this: EventListenerThisType<T>, ev: Event): any;
+    onCanPlayThroughCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onChange?(this: EventListenerThisType<T>, ev: Event): any;
+    onChangeCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onClick?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onClickCapture?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onCompassNeedsCalibration?(this: EventListenerThisType<T>, ev: Event): any;
+    onCompassNeedsCalibrationCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onContextMenu?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onContextMenuCapture?(this: EventListenerThisType<T>, ev: PointerEvent): any;
+    onDblClick?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onDblClickCapture?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onDeviceMotion?(this: EventListenerThisType<T>, ev: DeviceMotionEvent): any;
+    onDeviceMotionCapture?(this: EventListenerThisType<T>, ev: DeviceMotionEvent): any;
+    onDeviceOrientation?(this: EventListenerThisType<T>, ev: DeviceOrientationEvent): any;
+    onDeviceOrientationCapture?(this: EventListenerThisType<T>, ev: DeviceOrientationEvent): any;
+    onDrag?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDragCapture?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDragEnd?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDragEndCapture?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDragEnter?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDragEnterCapture?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDragLeave?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDragLeaveCapture?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDragOver?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDragOverCapture?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDragStart?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDragStartCapture?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDrop?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDropCapture?(this: EventListenerThisType<T>, ev: DragEvent): any;
+    onDurationChange?(this: EventListenerThisType<T>, ev: Event): any;
+    onDurationChangeCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onEmptied?(this: EventListenerThisType<T>, ev: Event): any;
+    onEmptiedCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onEnded?(this: EventListenerThisType<T>, ev: Event): any;
+    onEndedCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onError?(this: EventListenerThisType<T>, ...args: OnErrorEventHandlerArgs): any;
+    onErrorCapture?(this: EventListenerThisType<T>, ...args: OnErrorEventHandlerArgs): any;
+    onFocus?(this: EventListenerThisType<T>, ev: FocusEvent): any;
+    onFocusCapture?(this: EventListenerThisType<T>, ev: FocusEvent): any;
+    onHashChange?(this: EventListenerThisType<T>, ev: HashChangeEvent): any;
+    onHashChangeCapture?(this: EventListenerThisType<T>, ev: HashChangeEvent): any;
+    onInput?(this: EventListenerThisType<T>, ev: Event): any;
+    onInputCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onKeyDown?(this: EventListenerThisType<T>, ev: KeyboardEvent): any;
+    onKeyDownCapture?(this: EventListenerThisType<T>, ev: KeyboardEvent): any;
+    onKeyPress?(this: EventListenerThisType<T>, ev: KeyboardEvent): any;
+    onKeyPressCapture?(this: EventListenerThisType<T>, ev: KeyboardEvent): any;
+    onKeyUp?(this: EventListenerThisType<T>, ev: KeyboardEvent): any;
+    onKeyUpCapture?(this: EventListenerThisType<T>, ev: KeyboardEvent): any;
+    onLoad?(this: EventListenerThisType<T>, ev: Event): any;
+    onLoadCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onLoadedData?(this: EventListenerThisType<T>, ev: Event): any;
+    onLoadedDataCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onLoadedMetadata?(this: EventListenerThisType<T>, ev: Event): any;
+    onLoadedMetadataCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onLoadStart?(this: EventListenerThisType<T>, ev: Event): any;
+    onLoadStartCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onMessage?(this: EventListenerThisType<T>, ev: MessageEvent): any;
+    onMessageCapture?(this: EventListenerThisType<T>, ev: MessageEvent): any;
+    onMouseDown?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseDownCapture?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseEnter?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseEnterCapture?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseLeave?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseLeaveCapture?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseMove?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseMoveCapture?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseOut?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseOutCapture?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseOver?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseOverCapture?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseUp?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseUpCapture?(this: EventListenerThisType<T>, ev: MouseEvent): any;
+    onMouseWheel?(this: EventListenerThisType<T>, ev: WheelEvent): any;
+    onMouseWheelCapture?(this: EventListenerThisType<T>, ev: WheelEvent): any;
+    onMsGestureChange?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsGestureChangeCapture?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsGestureDoubleTap?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsGestureDoubleTapCapture?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsGestureEnd?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsGestureEndCapture?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsGestureHold?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsGestureHoldCapture?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsGestureStart?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsGestureStartCapture?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsGestureTap?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsGestureTapCapture?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsInertiaStart?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsInertiaStartCapture?(this: EventListenerThisType<T>, ev: MSGestureEvent): any;
+    onMsPointerCancel?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerCancelCapture?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerDown?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerDownCapture?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerEnter?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerEnterCapture?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerLeave?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerLeaveCapture?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerMove?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerMoveCapture?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerOut?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerOutCapture?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerOver?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onMsPointerOverCapture?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    oNmsPointerUp?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    oNmsPointerUpCapture?(this: EventListenerThisType<T>, ev: MSPointerEvent): any;
+    onOffline?(this: EventListenerThisType<T>, ev: Event): any;
+    onOfflineCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onOnline?(this: EventListenerThisType<T>, ev: Event): any;
+    onOnlineCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onOrientationChange?(this: EventListenerThisType<T>, ev: Event): any;
+    onOrientationChangeCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onPageHide?(this: EventListenerThisType<T>, ev: PageTransitionEvent): any;
+    onPageHideCapture?(this: EventListenerThisType<T>, ev: PageTransitionEvent): any;
+    onPageShow?(this: EventListenerThisType<T>, ev: PageTransitionEvent): any;
+    onPageShowCapture?(this: EventListenerThisType<T>, ev: PageTransitionEvent): any;
+    onPause?(this: EventListenerThisType<T>, ev: Event): any;
+    onPauseCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onPlay?(this: EventListenerThisType<T>, ev: Event): any;
+    onPlayCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onPlaying?(this: EventListenerThisType<T>, ev: Event): any;
+    onPlayingCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onPopState?(this: EventListenerThisType<T>, ev: PopStateEvent): any;
+    onPopStateCapture?(this: EventListenerThisType<T>, ev: PopStateEvent): any;
+    onProgress?(this: EventListenerThisType<T>, ev: ProgressEvent): any;
+    onProgressCapture?(this: EventListenerThisType<T>, ev: ProgressEvent): any;
+    onRateChange?(this: EventListenerThisType<T>, ev: Event): any;
+    onRateChangeCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onReadyStateChange?(this: EventListenerThisType<T>, ev: ProgressEvent): any;
+    onReadyStateChangeCapture?(this: EventListenerThisType<T>, ev: ProgressEvent): any;
+    onReset?(this: EventListenerThisType<T>, ev: Event): any;
+    onResetCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onResize?(this: EventListenerThisType<T>, ev: UIEvent): any;
+    onResizeCapture?(this: EventListenerThisType<T>, ev: UIEvent): any;
+    onScroll?(this: EventListenerThisType<T>, ev: UIEvent): any;
+    onScrollCapture?(this: EventListenerThisType<T>, ev: UIEvent): any;
+    onSeeked?(this: EventListenerThisType<T>, ev: Event): any;
+    onSeekedCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onSeeking?(this: EventListenerThisType<T>, ev: Event): any;
+    onSeekingCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onSelect?(this: EventListenerThisType<T>, ev: UIEvent): any;
+    onSelectCapture?(this: EventListenerThisType<T>, ev: UIEvent): any;
+    onSelectionChange?(this: EventListenerThisType<T>, ev: Event): any;
+    onSelectionChangeCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onStalled?(this: EventListenerThisType<T>, ev: Event): any;
+    onStalledCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onStorage?(this: EventListenerThisType<T>, ev: StorageEvent): any;
+    onStorageCapture?(this: EventListenerThisType<T>, ev: StorageEvent): any;
+    onSubmit?(this: EventListenerThisType<T>, ev: Event): any;
+    onSubmitCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onSuspend?(this: EventListenerThisType<T>, ev: Event): any;
+    onSuspendCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onTimeUpdate?(this: EventListenerThisType<T>, ev: Event): any;
+    onTimeUpdateCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onTouchCancel?(this: EventListenerThisType<T>, ev: TouchEvent): any;
+    onTouchCancelCapture?(this: EventListenerThisType<T>, ev: TouchEvent): any;
+    onTouchEnd?(this: EventListenerThisType<T>, ev: TouchEvent): any;
+    onTouchEndCapture?(this: EventListenerThisType<T>, ev: TouchEvent): any;
+    onTouchMove?(this: EventListenerThisType<T>, ev: TouchEvent): any;
+    onTouchMoveCapture?(this: EventListenerThisType<T>, ev: TouchEvent): any;
+    onTouchStart?(this: EventListenerThisType<T>, ev: TouchEvent): any;
+    onTouchStartCapture?(this: EventListenerThisType<T>, ev: TouchEvent): any;
+    onUnload?(this: EventListenerThisType<T>, ev: Event): any;
+    onUnloadCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onVolumeChange?(this: EventListenerThisType<T>, ev: Event): any;
+    onVolumeChangeCapture?(this: EventListenerThisType<T>, ev: Event): any;
+    onWaiting?(this: EventListenerThisType<T>, ev: Event): any;
+    onWaitingCapture?(this: EventListenerThisType<T>, ev: Event): any;
 
     /**
      * Target (window or document)
      */
-    target: Window | Document | string;
+    target: T;
 }
 
-export default class EventListener extends React.Component<EventListenerProps> { }
+export type WindowEventTargets = {[K in keyof Window]: Window[K] extends EventTarget ? K : never}[keyof Window];
+
+export default class EventListener<T extends EventTarget | WindowEventTargets> extends React.PureComponent<EventListenerProps<T>> { }
