--- types/react-event-listener/package.json
@@ -1,11 +1,0 @@
-{
-  "private": true,
-  "types": "index",
-  "typesVersions": {
-    ">=3.1.0-0": {
-      "*": [
-        "ts3.1/*"
-      ]
-    }
-  }
-}
