+++ types/react-event-listener/react-event-listener-tests.tsx
@@ -5,3 +5,5 @@ import EventListener, { withOptions } from "react-event-listener";
 <EventListener target={window} onResize={withOptions((ev: UIEvent) => { }, { passive: true, capture: true })} />;
 <EventListener target="window" onResize={() => { }}/>;
 <EventListener target={document} onSelectionChange={ev => { }} />;
+<EventListener target="document" onError={function() { this.documentElement; }} />;
+<EventListener target={window} onBlurCapture={function() { this.pageXOffset; }} />;
